Many tests failed because they the contacts need to be in the contact list other wise it throws an illegalArgumentException.
Inserted addContacts() into @Before createManager() to add the contacts
in before every test. Made small changes to every time a contact is
created so that they match with the initial list of contacts.
Finaly, before every time addContacts() was already used, created a new
instance of contactManager to ensure that the list of contacts is not
doubled.